Sletter kladd når man reaktiverer deltakelser fordi det fører nå til problemer med at man kan videreføre kladden etter at man har reaktivert deltakelse som fører til flere aktive deltakelser på personen
